+++ CHANGELOG.md
@@ -105,6 +105,8 @@ Please also have a look at our
 
 ### Fixed
 
+- Insert `Rule` before sibling even with different property name
+  (in `RuleSet::addRule()`) (#1270)
 - Ensure `RuleSet::addRule()` sets non-negative column number when sibling
   provided (#1268)
 - Set line number when `RuleSet::addRule()` called with only column number set